refactor: replace Sidebar with Navbar for improved navigation and layout consistency
