+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "study-pal",
   "private": true,
-  "version": "0.0.3",
+  "version": "0.3.1",
   "type": "module",
   "scripts": {
     "dev": "vite",
